Civic_issue-main/Civic_issue-main/backend/requirements.txt
Pillow==11.2.1
exifread==3.0.0
ImageHash==4.3.1
ultralytics>=8.0.0
torch>=2.0.0
torchvision>=0.15.0

